+++ Client/src/pages/profile.js
@@ -6,7 +6,7 @@ import './Basic.css';
 //  - ADDITIONAL PLAYER RENDER HERE
 
 // {MAIN}/profile/:profileName
-export const ProfileBase = (props) => {
+export const profileBase = (props) => {
     const profileName = props.match.params.profileName;
 
     return (
@@ -17,7 +17,7 @@ export const ProfileBase = (props) => {
 };
 
 // {MAIN}/profile/:profileName/champs
-export const ProfileChamps = (props) => {
+export const profileChamps = (props) => {
     const profileName = props.match.params.profileName;
 
     return (
@@ -28,7 +28,7 @@ export const ProfileChamps = (props) => {
 };
 
 // {MAIN}/profile/:profileName/games/:seasonShortName
-export const ProfileGames = (props) => {
+export const profileGames = (props) => {
     const profileName = props.match.params.profileName;
     const seasonName = props.match.params.seasonShortName;
 
@@ -40,7 +40,7 @@ export const ProfileGames = (props) => {
 };
 
 // {MAIN}/profile/:profileName/stats/:tournamentShortName
-export const ProfileStats = (props) => {
+export const profileStats = (props) => {
     const profileName = props.match.params.profileName;
     const tourneyName = props.match.params.tournamentShortName;